Added the ability to get named channels

+++ aurochs-core/src/main/java/se/l4/aurochs/core/io/ByteMessage.java
@@ -4,7 +4,7 @@ import java.util.function.Predicate;
 
 public interface ByteMessage
 {
-	int getTag();
+	long getTag();
 	
 	Bytes getData();
